+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 0.1.0
 -----
 
+- Add `deploy:teardown` to remove the application from remote servers
 - Add support for the connections defined in the remote config file
 - Add `deploy:rollback` and `deploy:current` commands
 - Add `deploy:cleanup` command